Create a shared google sheet module
This will be used in multiple places so extracting
